Add folder selector component
Adds a file selector which allows the user to select any folder.

While the API returns a list of paths of individual files,
the components returns the folder that was selected.

This matches the setup of the analyzer and will allow us
to use a glob pattern to select only audio files.