Add sound list with audio waveform + styling changes
Instead of using the native web audio player, `wavesurfer.js` is now
added which shows a waveform of the audio as it plays.
In the future we can use this for cool visualizations, see http://wavesurfer-js.org/examples/.

This commit, additionally re-works the UI to be a minimal white design with
blue accents.  To support this `react-icons` is added to give more
design assets which power the audio player and new header.

The UI now initially shows cards which can
be used to select a classification/sorting type.
Then a split list will show each audio file categorized by likeness on the sort type.
(We could also add a more graphical view here later on as a tab.)