Improve audio player & slider behavior (#17)
This commit addresses a few things to make audio playback feel and sound better.

The first of which is it fixes a bug where multiple samples can be played at the same time.  This was a mistake in the audio pause state reducer.

Additionally, the audio slider is now properly debounced and will only skip ahead once the slider is released.	Prior to this it would
update whenever the mouse button was held leading to weird audio artifacts.

Also pressing the spacebar will now pause/play the selected audio track.