Prepare app for Windows packaging (#37)
Prepares the app for Windows packaging.  Prior to this commit, a few things prevented the app from running on Windows:

- The separators in paths were not normalized by the folder selector to use the correct separator.
- Glob only supports using the forward slash so using a Windows path returned 0 results.  I was able to work around this since it allows Windows paths as the root.
- The custom `audio://` protocol normalized the Windows path to remove colons and backslashes.	This was fixed by making it a non standard protocol.
- The analyzer interestingly returns 0 results for the `tasks` field which caused the progress bar to not show up.  I worked around this by using the latest field instead

To package it the database needed to be stored outside the asar app package.  Additionally the worker needed to be included in this package.

I tested these changes on both Debian and Windows it everything seems to work the same.